feat: modify example and fix things
