+++ example/src/components/index.js
@@ -1,4 +1,4 @@
-export { default as SimpleClientNode } from "./SimpleClientNode";
-export { default as SimplePublisherNode } from "./SimplePublisherNode";
-export { default as SimpleServiceNode } from "./SimpleServiceNode";
-export { default as SimpleSubscriptionNode } from "./SimpleSubscriptionNode";
+export { default as SimpleClient } from "./SimpleClient";
+export { default as SimplePublisher } from "./SimplePublisher";
+export { default as SimpleService } from "./SimpleService";
+export { default as SimpleSubscription } from "./SimpleSubscription";
