+++ package.json
@@ -29,7 +29,7 @@
   },
   "dependencies": {
     "@material-ui/core": "^4.11.3",
-    "kumo-client": "^0.2.1",
+    "kumo-client": "^0.2.3",
     "notistack": "^1.0.5",
     "prop-types": "^15.7.2",
     "store2": "^2.12.0"
